// src/services/usos/types.ts
export enum Frequency {
  EVERY = "każd",
  ODD = "nieparzyst",
  EVEN = "parzyst",
  NEVER = "NEVER",
}

export enum Day {
  MONDAY = "poniedzi",
  TUESDAY = "wtor",
  WEDNESDAY = "środ",
  THURSDAY = "czwart",
  FRIDAY = "piąt",
  SATURDAY = "sobot",
  SUNDAY = "niedziel",
  NEVER = "NEVER",
}

export enum LessonType {
  LECTURE = "wyklad",
  EXERCISES = "cwiczenia",
  LABORATORY = "laboratorium",
  PROJECT = "projekt",
  SEMINAR = "seminarium",
}